requirements.txt
google-api-python-client
python-dotenv
pandas
numpy
scikit-learn
xgboost
joblib
shap
matplotlib
torch
beautifulsoup4
